Split test_enchant across several files
Note: we also removed tests about py2exe and the code that used
to run the whole test suite twice in different folders (once of
them containing non-ASCII characters)

We'll see if wee need to re-enable them somehow
